Refactor: 마이 대쉬보드 목록 불러오기 tanstack query로 리팩토링
(cherry picked from commit 65ccd2c4599144757524a883cb3baaa828432fe7)
